+++ src/main/java/be/ac/umons/Class/PizzaComponent.java
@@ -2,8 +2,8 @@ package be.ac.umons.Class;
 
 public interface PizzaComponent {
     public abstract String getName();
-    public abstract float getPrice();
+    public abstract double getPrice();
     public abstract void setName(String name);
-    public abstract void setPrice(float price);
+    public abstract void setPrice(double price);
     public abstract String toString();
 }
